# Improve preset listing performance