Eliminating the copy accelerates the program listing by a vast amount.

In DSSI, performance is a problem with naspro-bridges installed, because when it lists program in a loop as the host loads LV2, the host will hang for a whole minute while it processes.

Before: `./load-amsynth .libs/amsynth_dssi.so  58,52s user 0,19s system 98% cpu 59,699 total`
After: `./load-amsynth .libs/amsynth_dssi.so  0,37s user 0,05s system 98% cpu 0,421 total`

I've applied some identical changes throughout the program.

[load-amsynth.cc.gz](https://github.com/amsynth/amsynth/files/3236815/load-amsynth.cc.gz)